src, tests: add support for other GeoJSON features in toArray()
fixes #2. It doesn't make sense to support `toArray()` for a `GeometryCollection`, `Feature` or `FeatureCollection` because they are intended to carry more attributes, including other primary `GeoJSON` objects.
The unit tests added are ones that check that a valid array is returned, will add tests to check for invalidations over time. This has been long outstanding.